+++ CHANGES.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## 0.2.0 WIP
+- Import other trove files `hoard import --file /path/to/trove.yml`
+- Import trove files from url `hoard import --url https://this.trove.com/trove.yml`
+
 ## 0.1.4
 
 - Can now delete all commands in a specific namespace `hoard remove --namespace <name>`